requirements.txt
Flask==2.3.2
python-telegram-bot==20.7
openpyxl==3.1.2
xlrd==2.0.1
rapidfuzz==3.9.1
gunicorn==21.2.0
python-dotenv
